Possibility of incorrect image used
"The recipe above has a trigger and just one action. The trigger is **New Salesforce account** - this trigger will fire whenever a new account is created in Salesforce. The action **Create Zendesk organization** will create an organization in Zendesk each time the trigger event occurs (i.e. when a new account is created in Salesforce)." => does not correspond to the image recipe-id.png, but instead corresponds to the image input-fields.png.